moved cargo_metadata to syft/pkg
